Reorganize directory structure, place layout components under app
