+++ src/app/(main)/our-portfolio/page.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/components/Navbar";
+import Navbar from "@/app/components/navbar/Navbar";
 import Image from "next/image";
 import Link from "next/link";
