+++ src/app/components/ContactForm.tsx
@@ -1,11 +1,11 @@
 "use client";
 // src/components/ContactForm.tsx
 import { useServiceTheme } from "@/app/incore/services/[service]/hooks/useServiceTheme"; // added
-import { Container } from "@/components/ui/container";
+import { Container } from "@/components/container";
 import { MapPin } from "lucide-react";
-import { Button } from "./ui/button";
-import { Input } from "./ui/input";
-import { Textarea } from "./ui/textarea";
+import { Button } from "../../components/ui/button";
+import { Input } from "../../components/ui/input";
+import { Textarea } from "../../components/ui/textarea";
 
 export const ContactForm = () => {
   const theme = useServiceTheme();
