+++ src/app/components/Footer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useServiceTheme } from "@/app/incore/services/[service]/hooks/useServiceTheme";
-import { Container } from "@/components/ui/container";
+import { Container } from "@/components/container";
 
 export const Footer = () => {
   const theme = useServiceTheme();
