+++ src/app/components/navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 ﻿"use client";
 
+import { getNavConfig, type NavItem } from "@/app/components/navbar/navigation";
 import { Button } from "@/components/ui/button";
-import { getNavConfig, type NavItem } from "@/config/navigation";
 import { ChevronRight, Menu, Search, X } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
