+++ src/app/incore/components/DecorativeBannerSection.tsx
@@ -1,4 +1,4 @@
-import Section from "@/components/ui/section";
+import Section from "@/components/section";
 import { cn } from "@/lib/utils";
 import Image from "next/image";
 import Link from "next/link";
