+++ src/app/incore/layout.tsx
@@ -1,9 +1,9 @@
-import Navbar from "@/components/Navbar";
-import TopBanner from "@/components/TopBanner";
+import Navbar from "@/app/components/navbar/Navbar";
+import TopBanner from "@/app/components/TopBanner";
 import type { Metadata } from "next";
 import Image from "next/image";
 import Link from "next/link";
-import { Footer } from "../(main)/(home)/Footer";
+import { Footer } from "../components/Footer";
 
 export const metadata: Metadata = {
   title: {
