+++ src/app/incore/services/[service]/coreOfferings.tsx
@@ -2,8 +2,8 @@
 
 import { useServiceContent } from "@/app/incore/services/[service]/hooks/useServiceContent";
 import { useServiceTheme } from "@/app/incore/services/[service]/hooks/useServiceTheme";
+import Container from "@/components/container";
 import { Button } from "@/components/ui/button";
-import Container from "@/components/ui/container";
 import Cube from "@public/inCore/cube.svg";
 import CircledLine from "@public/inCore/text-circled-line.svg";
 import { ArrowRight } from "lucide-react";
