--- src/styles/tw-safelist.ts
@@ -1,2 +1,0 @@
-// (Removed) Previously used for dynamic text opacity safelist. No longer needed after reverting theme accessor.
-export {};
